+++ gateway/src/backend/start.rs
@@ -1,9 +1,9 @@
-use crate::backend_commands::BackendCommands; 
+use crate::backend_commands::BackendCommands;
 use crate::kafka_io;
 use crate::redis_io;
 
-use crossbeam_channel::{   unbounded, Receiver, Sender};
-use futures::executor::block_on; 
+use crossbeam_channel::{unbounded, Receiver, Sender};
+use futures::executor::block_on;
 use std::thread;
 
 use super::*;
@@ -28,7 +28,7 @@ pub fn start_all(opts: BackendInitOptions) {
 
     let soc = opts.session_commands_out;
     thread::spawn(move || {
-        split_commands(SplitOpts {
+        split_commands(super::split::SplitOpts {
             session_commands_out: soc,
             kafka_commands_in,
             redis_commands_in,
@@ -44,4 +44,11 @@ pub fn start_all(opts: BackendInitOptions) {
         opts.kafka_activity_in.clone(),
         kafka_commands_out,
     ))
+}
+
+pub struct BackendInitOptions {
+    pub backend_events_in: Sender<BackendEvents>,
+    pub shutdown_in: Sender<KafkaShutdownEvent>,
+    pub kafka_activity_in: Sender<KafkaActivityObserved>,
+    pub session_commands_out: Receiver<SessionCommands>,
 }
